requirements.txt
arrow==1.3.0
asgiref==3.7.2
binaryornot==0.4.4
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.5.0
Django==5.0.1
django-browser-reload==1.12.1
django-filter==23.5
django-tailwind==3.8.0
idna==3.6
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mdurl==0.1.2
pillow==10.2.0
Pygments==2.17.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-slugify==8.0.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
types-python-dateutil==2.8.19.20240106
tzdata==2023.4
urllib3==2.1.0
